% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeComparativeCohortAnalysis.R
\name{executeComparativeCohortAnalysis}
\alias{executeComparativeCohortAnalysis}
\title{Comparative Cohort Analysis}
\usage{
executeComparativeCohortAnalysis(treatmentId, comparatorId, outcomeId,
  modelType, executionId, dbms, connectionString, cdmTableQualifier,
  resultsTableQualifier, timeAtRiskStart, timeAtRiskEnd, addExposureDaysToEnd,
  minimumWashoutPeriod, minimumDaysAtRisk, rmSubjectsInBothCohorts,
  rmPriorOutcomes, psAdjustment, psExclusions = c(), psInclusions = c(),
  psDemographics, psDemographicsGender, psDemographicsRace,
  psDemographicsEthnicity, psDemographicsAge, psDemographicsYear,
  psDemographicsMonth, psTrim, psTrimFraction, psMatch, psMatchMaxRatio,
  psStrat, psStratNumStrata, psConditionOcc, psConditionOcc365d,
  psConditionOcc30d, psConditionOccInpt180d, psConditionEra, psConditionEraEver,
  psConditionEraOverlap, psConditionGroup, psConditionGroupMeddra,
  psConditionGroupSnomed, psDrugExposure, psDrugExposure365d, psDrugExposure30d,
  psDrugEra, psDrugEra365d, psDrugEra30d, psDrugEraOverlap, psDrugEraEver,
  psDrugGroup, psProcedureOcc, psProcedureOcc365d, psProcedureOcc30d,
  psProcedureGroup, psObservation, psObservation365d, psObservation30d,
  psObservationCount365d, psMeasurement, psMeasurement365d, psMeasurement30d,
  psMeasurementCount365d, psMeasurementBelow, psMeasurementAbove,
  psConceptCounts, psRiskScores, psRiskScoresCharlson, psRiskScoresDcsi,
  psRiskScoresChads2, psRiskScoresChads2vasc, psInteractionYear,
  psInteractionMonth, omCovariates, omExclusions = c(), omInclusions = c(),
  omDemographics, omDemographicsGender, omDemographicsRace,
  omDemographicsEthnicity, omDemographicsAge, omDemographicsYear,
  omDemographicsMonth, omTrim, omTrimFraction, omMatch, omMatchMaxRatio,
  omStrat, omStratNumStrata, omConditionOcc, omConditionOcc365d,
  omConditionOcc30d, omConditionOccInpt180d, omConditionEra, omConditionEraEver,
  omConditionEraOverlap, omConditionGroup, omConditionGroupMeddra,
  omConditionGroupSnomed, omDrugExposure, omDrugExposure365d, omDrugExposure30d,
  omDrugEra, omDrugEra365d, omDrugEra30d, omDrugEraOverlap, omDrugEraEver,
  omDrugGroup, omProcedureOcc, omProcedureOcc365d, omProcedureOcc30d,
  omProcedureGroup, omObservation, omObservation365d, omObservation30d,
  omObservationCount365d, omMeasurement, omMeasurement365d, omMeasurement30d,
  omMeasurementCount365d, omMeasurementBelow, omMeasurementAbove,
  omConceptCounts, omRiskScores, omRiskScoresCharlson, omRiskScoresDcsi,
  omRiskScoresChads2, omRiskScoresChads2vasc, omInteractionYear,
  omInteractionMonth, delCovariatesSmallCount, negativeControls)
}
\arguments{
\item{treatment}{The treatment cohort definition identifier}

\item{comparator}{The comparator cohort definition identifier}

\item{outcome}{The outcome cohort definition identifier}
}
\description{
Comparative Cohort Analysis
}
\keyword{cohort}
\keyword{comparative}

